Move semantic-specific stuff back to Data.Language.
