+++ semantic-analysis/src/Analysis/Language.hs
@@ -11,13 +11,8 @@ module Analysis.Language
   , knownLanguage
   , forPath
   , pathIsMinified
-  , supportedExts
-  , codeNavLanguages
   , textToLanguage
   , languageToText
-  , PerLanguageModes(..)
-  , defaultLanguageModes
-  , LanguageMode(..)
   ) where
 
 import           Data.Aeson
@@ -109,17 +104,6 @@ forPath path =
     [result] -> textToLanguage result
     _        -> Unknown
 
-supportedExts :: [String]
-supportedExts = foldr append mempty supportedLanguages
-  where
-    append (Just l) b = fmap T.unpack (Lingo.languageExtensions l) <> b
-    append Nothing  b = b
-    supportedLanguages = fmap lookup (languageToText <$> codeNavLanguages)
-    lookup k = Map.lookup k Lingo.languages
-
-codeNavLanguages :: [Language]
-codeNavLanguages = [Go, Java, Ruby, Python, JavaScript, TypeScript, PHP]
-
 pathIsMinified :: FilePath -> Bool
 pathIsMinified = OldPath.isExtensionOf ".min.js"
 
@@ -156,29 +140,3 @@ textToLanguage = \case
   _ -> Unknown
 
 
-data PerLanguageModes = PerLanguageModes
-  { pythonMode     :: LanguageMode
-  , rubyMode       :: LanguageMode
-  , goMode         :: LanguageMode
-  , typescriptMode :: LanguageMode
-  , tsxMode        :: LanguageMode
-  , javascriptMode :: LanguageMode
-  , jsxMode        :: LanguageMode
-  }
-  deriving (Eq, Ord, Show)
-
-defaultLanguageModes :: PerLanguageModes
-defaultLanguageModes = PerLanguageModes
-  { pythonMode = ALaCarte
-  , rubyMode = ALaCarte
-  , goMode = ALaCarte
-  , typescriptMode = ALaCarte
-  , tsxMode = ALaCarte
-  , javascriptMode = ALaCarte
-  , jsxMode = ALaCarte
-  }
-
-data LanguageMode
-  = ALaCarte
-  | Precise
-  deriving (Bounded, Enum, Eq, Ord, Read, Show)
